// src/types/partials/HumanName.ts
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TPeriod } from '../partials/Period';

export type THumanName = {
    id?: String;
    extension?: TExtension[];
    use?: String;
    text?: String;
    family?: String;
    given?: String[];
    prefix?: String[];
    suffix?: String[];
    period?: TPeriod;
};